🔨Use double? instead of int? in FoodDto

+++ life-manager-backend/Models/FoodDto.cs
@@ -4,15 +4,15 @@
     {
         public long Id { get; set; }
         public string Name { get; set; } = String.Empty;
-        public int? Energy { get; set; }
-        public int? Fat { get; set; }
-        public int? SaturatedFat { get; set; }
-        public int? MonoUnsaturatedFat { get; set; }
-        public int? PolyUnsaturatedFat { get; set; }
-        public int? Carbohydrates { get; set; }
-        public int? Sugars { get; set; }
-        public int? Protein { get; set; }
-        public int? Salt { get; set; }
-        public int? Fiber { get; set; }
+        public double? Energy { get; set; }
+        public double? Fat { get; set; }
+        public double? SaturatedFat { get; set; }
+        public double? MonoUnsaturatedFat { get; set; }
+        public double? PolyUnsaturatedFat { get; set; }
+        public double? Carbohydrates { get; set; }
+        public double? Sugars { get; set; }
+        public double? Protein { get; set; }
+        public double? Salt { get; set; }
+        public double? Fiber { get; set; }
     }
 }
